# Conflicts:
#	Back-End/src/main/java/com/conseller/conseller/barter/barter/barterService/BarterServiceImpl.java
#	Back-End/src/main/java/com/conseller/conseller/user/UserController.java
#	Back-End/src/main/java/com/conseller/conseller/user/service/UserService.java
#	Back-End/src/main/java/com/conseller/conseller/user/service/UserServiceImpl.java

+++ Back-End/src/main/java/com/conseller/conseller/entity/User.java
@@ -47,7 +47,7 @@ public class User extends BaseTime implements UserDetails {
     private Integer userAge;
 
     @Column(name = "user_deposit", nullable = false)
-    private Integer userDeposit;
+    private Long userDeposit;
 
     @Column(name = "user_deleted_date")
     private LocalDateTime userDeletedDate;
