+++ Back-End/src/main/java/com/conseller/conseller/user/service/UserServiceImpl.java
@@ -7,6 +7,7 @@ import com.conseller.conseller.auction.bid.dto.mapper.AuctionBidMapper;
 import com.conseller.conseller.auction.bid.dto.response.AuctionBidResponse;
 import com.conseller.conseller.barter.barter.barterDto.response.BarterResponseDto;
 import com.conseller.conseller.barter.barter.barterDto.response.MyBarterResponseDto;
+import com.conseller.conseller.barter.barter.barterDto.response.BarterResponseDTO;
 import com.conseller.conseller.barter.barterRequest.barterRequestDto.MyBarterRequestResponseDto;
 import com.conseller.conseller.entity.*;
 import com.conseller.conseller.gifticon.dto.response.GifticonResponse;
@@ -62,7 +63,7 @@ public class UserServiceImpl implements UserService {
                 .userId(signUpRequest.getUserId())
                 .userPassword(signUpRequest.getUserPassword())
                 .userEmail(signUpRequest.getUserEmail())
-                .userDeposit(0)
+                .userDeposit((long) 0)
                 .userNickname(signUpRequest.getUserNickname())
                 .userPhoneNumber(signUpRequest.getUserPhoneNumber())
                 .userGender(signUpRequest.getUserGender())
@@ -213,7 +214,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public void deposit(long userIdx, int deposit) {
+    public void deposit(long userIdx, Long deposit) {
         User user = userRepository.findByUserIdx(userIdx)
                 .orElseThrow(() -> new RuntimeException("없는 유저 입니다."));
         user.setUserDeposit(deposit);
